added missing fields and registers

+++ src/usci_a0_uart_mode/uca0txbuf.rs
@@ -37,7 +37,7 @@ impl From<crate::W<UCA0TXBUF_SPEC>> for W {
 impl W {
     #[doc = "Writes raw bits to the register."]
     #[inline(always)]
-    pub unsafe fn bits(&mut self, bits: u16) -> &mut Self {
+    pub unsafe fn bits(&mut self, bits: u8) -> &mut Self {
         self.0.bits(bits);
         self
     }
@@ -45,7 +45,7 @@ impl W {
 #[doc = "USCI A0 Transmit Buffer\n\nThis register you can [`read`](crate::generic::Reg::read), [`write_with_zero`](crate::generic::Reg::write_with_zero), [`reset`](crate::generic::Reg::reset), [`write`](crate::generic::Reg::write), [`modify`](crate::generic::Reg::modify). See [API](https://docs.rs/svd2rust/#read--modify--write-api).\n\nFor information about available fields see [uca0txbuf](index.html) module"]
 pub struct UCA0TXBUF_SPEC;
 impl crate::RegisterSpec for UCA0TXBUF_SPEC {
-    type Ux = u16;
+    type Ux = u8;
 }
 #[doc = "`read()` method returns [uca0txbuf::R](R) reader structure"]
 impl crate::Readable for UCA0TXBUF_SPEC {
